chore: split solution to separate files
